Refactor security pool state helper exports
- Split lifecycle labels and axis derivation into dedicated modules
- Update components and tests to use the consolidated state API
