--- ui/ts/lib/securityPoolState/index.ts
@@ -1,7 +1,0 @@
-export { evaluateSecurityPoolState } from './engine.js'
-export { getSecurityPoolLifecycleLabel } from './labels.js'
-export type {
-	SecurityPoolActionId,
-	SecurityPoolLifecycleState,
-	SecurityPoolStateModel,
-} from './types.js'
